Refining Layout + Minor Fixes

+++ app/src/main/java/com/fadergs/victorhaps/Jogador.java
@@ -29,7 +29,7 @@ public class Jogador {
         this.nome = nome;
     }
 
-    public double getCamisa() {
+    public int getCamisa() {
         return camisa;
     }
